# api_gateway_service/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a REST API service for creating short urls.",
        "title": "Url Shortener",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:10000",
    "basePath": "/.",
    "paths": {
        "/short": {
            "post": {
                "description": "Create short url from url and return it",
                "tags": [
                    "ShortURL"
                ],
                "summary": "Create short url",
                "operationId": "create-short-url",
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            }
        },
        "/u/{token}": {
            "get": {
                "description": "Redirect user from short url to original url",
                "tags": [
                    "ShortURL"
                ],
                "summary": "Redirect to original url",
                "operationId": "redirect-short-url",
                "responses": {
                    "303": {
                        "description": "See Other"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            }
        },
        "/users": {
            "post": {
                "description": "Create user",
                "tags": [
                    "User"
                ],
                "summary": "Create user",
                "operationId": "create-user",
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            }
        },
        "/users/{id}": {
            "get": {
                "description": "Get user",
                "tags": [
                    "User"
                ],
                "summary": "Get user",
                "operationId": "get-user",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            },
            "delete": {
                "description": "Delete user",
                "tags": [
                    "User"
                ],
                "summary": "Delete user",
                "operationId": "delete-user",
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            }
        }
    }
}